requirements.txt
fastapi>=0.135.0
uvicorn>=0.42.0
minio>=7.2.0
psycopg2-binary>=2.9.0
pandas>=3.0.0
huggingface_hub>=1.9.0
python-dotenv>=1.2.0
pyyaml>=6.0.0
